fix(protocol): keep the Android v18 record_index in the unsigned domain
`record_index@11` is an unsigned 32-bit field. The Kotlin decoder narrowed it
with `toInt()`, and Kotlin's Int is 32-bit where Swift's is 64-bit, so a value
with bit 31 set decoded to -1_062_772_323 on Android and 3_232_194_973 on
Swift/macOS from byte-identical bytes.

The per-platform fixture-hex tests cannot see this: the wire bytes are the same
on both platforms and each suite asserts its own platform's answer, so the
divergence lives entirely downstream of the bytes. PR #848 hit the same
32-vs-64-bit split in the v18 storage codec and had to re-widen there; this is
the decoder key itself.

Carried as a Long, matching the u32 reader's own return type. No main-code
consumer reads the key (the only readers are tests and the field census), so
the change is confined to the decoded map and its assertion.

+++ android/app/src/main/java/com/noop/protocol/HistoricalStreams.kt
@@ -313,7 +313,13 @@ private fun decodeWhoop5Historical(frame: ByteArray): Map<String, Any?>? {
     out["hist_version"] = version
     // @11 a per-record counter: +1 every record, independent of unix (advances across gaps); seen on
     // two straps. @11 is only the low byte — read the full u32 LE.
-    frame.histU32(11)?.let { out["record_index"] = it.toInt() }
+    // Emitted as a Long, NOT narrowed to Int: this is an UNSIGNED 32-bit field and Kotlin's Int is
+    // 32-bit while Swift's is 64-bit, so `toInt()` made a value with bit 31 set decode to
+    // -1_062_772_323 here and 3_232_194_973 in Swift from byte-identical bytes — a cross-platform
+    // decode divergence the per-platform fixture-hex tests cannot see (same bytes, each suite asserting
+    // its own answer). PR #848 hit the same 32-vs-64-bit split in the storage codec. Pinned on both
+    // platforms by the `whoop5_v18_synthetic_record_index_high_bit` fixture in decoder_oracle.json.
+    frame.histU32(11)?.let { out["record_index"] = it }
     frame.histU32(15)?.let { out["unix"] = it.toInt() }
     frame.histU8(22)?.let { out["heart_rate"] = it }
     val rrn = frame.histU8(23) ?: 0
